v3 review-2026-05-05 §1.1 §2.7 §5: mechanical cleanups
§1.1: OP_RETURN's local fFlags var was uint8_t copied from CallFrame::
flags (uint32_t).  Latent today (both bits fit in 8) but the comments
anticipate more flags.  One-token fix.

§2.7: disasm.cc opName / opExtraWords were missing ~15 opcodes plus
the OP_CALL_PRIMOP extra-word (which dis-aligned every disassembly
past a primop call).  Add them all; document the operand-only ops in
a comment.

§5: dead env-var doc references (NIX_V3_RETURN_CHAIN, NIX_V3_OUTER_WITH,
NIX_V3_ON_DEMAND_ROOT_NEVER_RUN_OD).  These were doc-only mentions
with no corresponding getenv -- prune the comments to point at the
flags that actually exist (or to the simpler description).

+++ src/libexpr-v3/vm.cc
@@ -1554,7 +1554,7 @@ Value dispatchLoop(VMState & vm, size_t exitDepth)
             const CallFrame & frRef = vm.frames.back();
             const uint32_t fStackBase    = frRef.stackBaseOffset;
             const uint32_t fWithBase     = frRef.withStackBase;
-            const uint8_t  fFlags        = frRef.flags;
+            const uint32_t fFlags        = frRef.flags;
             Thunk *        fThunk        = frRef.thunk;
             vm.valueStack.resize(fStackBase);
             vm.withStack.resize(fWithBase);
@@ -1704,15 +1704,12 @@ Value dispatchLoop(VMState & vm, size_t exitDepth)
                 fr.thunk->state = ThunkState::Evaluated;
                 fr.thunk->evaluated = retVal;
 
-                // WC-38: the legacy "return-chain push" — eagerly
+                // WC-38: the legacy "return-chain push" -- eagerly
                 // forcing the next thunk if the outer's body returned
-                // a Suspended thunk — was kept behind
-                // NIX_V3_RETURN_CHAIN=1 for A/B testing after the
-                // WC-38 fix landed.  Phase-13 review MED-19 confirmed
-                // it's never needed: forceValue's chase loop already
-                // resolves the chain on the consumer's pull, and
-                // OP_RETURN's caller-resume path re-runs op_force_slow
-                // when CFF_FORCE_RETRY is set.  Removed entirely.
+                // a Suspended thunk -- has been removed.  forceValue's
+                // chase loop already resolves the chain on the
+                // consumer's pull, and OP_RETURN's caller-resume path
+                // re-runs op_force_slow when CFF_FORCE_RETRY is set.
             }
             if (vm.frames.size() == exitDepth) {
                 finalResult = retVal;
